+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.7.1] 2023-10-15
+
+- Fix [quiltdata/nf-quilt#145](https://github.com/quiltdata/nf-quilt/issues/145)
+- Remove automatic pre-install of packages
+
 ## [0.7.0] 2023-10-05
 
 - Officially QuiltCore 0.1.0 instead of Python